Revert "Fix warnings caused by casts between ints and pointers on 64bit"
This reverts commit 4293a258255849854670ce12de428581904a1a24.
The fix had introduced regressions on 32bit :<

+++ uzbl.h
@@ -27,7 +27,7 @@ enum {
 
 const struct {
     gchar *symbol_name;
-    gulong symbol_token;
+    guint symbol_token;
 } symbols[] = {
     {"NAME",                 SYM_NAME},
     {"URI",                  SYM_URI},
